requirements.txt
apache-flink==1.18.0
ibis-framework @ git+https://github.com/mfatihaktas/ibis.git@feat/temporal-join-2
kafka-python